+++ routes/web.php
@@ -3,7 +3,7 @@
 use App\Http\Controllers\Dashboard;
 use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\Settings;
-use App\Http\Controllers\UploadedFileController;
+use App\Http\Controllers\FileController;
 use App\Livewire\Development;
 use App\Livewire\Files;
 use App\Models\UserPermission;
@@ -21,8 +21,8 @@ use Illuminate\Support\Facades\Route;
 */
 
 Route::get('/', [Dashboard::class, 'index'])->name('home');
-Route::get('/file/{file}', [UploadedFileController::class, 'response'])->name('file');
-Route::get('/file/{file}/download', [UploadedFileController::class, 'download'])->name('file.download');
+Route::get('/file/{file}', [FileController::class, 'response'])->name('file');
+Route::get('/file/{file}/download', [FileController::class, 'download'])->name('file.download');
 Route::get('/dashboard', [Dashboard::class, 'index'])->name('dashboard');
 Route::get('/imprint', fn () => view('imprint'))->name('imprint');
 Route::get('/privacy-policy', fn () => view('privacy-policy'))->name('privacy-policy');