+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file.
 
 ## Unpublished
+### Fixed
+- no double separator lines on dashboard cards
 
 ## v0.7.16 - 2024-09-11
 ### Changed